+++ src/components/icon-button/icon-button.styles.ts
@@ -6,6 +6,7 @@ export default css`
 
   :host {
     display: inline-block;
+    color: var(--sl-color-neutral-600);
   }
 
   .icon-button {
@@ -16,7 +17,7 @@ export default css`
     border: none;
     border-radius: var(--sl-border-radius-medium);
     font-size: inherit;
-    color: var(--sl-color-neutral-600);
+    color: inherit;
     padding: var(--sl-spacing-x-small);
     cursor: pointer;
     transition: var(--sl-transition-medium) color;